refactor: dispatch github issue callbacks through ccstate
