+++ turbo/apps/api/src/signals/services/__tests__/agent-run-callback.service.test.ts
@@ -1,10 +1,13 @@
 import { createStore } from "ccstate";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { HttpResponse, http } from "msw";
 import { afterEach, describe, expect, it } from "vitest";
 import { agentRunCallbacks } from "@vm0/db/schema/agent-run-callback";
+import { agentSessions } from "@vm0/db/schema/agent-session";
 import { automations, automationTriggers } from "@vm0/db/schema/automation";
 import { chatThreads } from "@vm0/db/schema/chat-thread";
+import { githubInstallations } from "@vm0/db/schema/github-installation";
+import { githubIssueSessions } from "@vm0/db/schema/github-issue-session";
 import { vm0ApiKeys } from "@vm0/db/schema/vm0-api-key";
 import { zeroRuns } from "@vm0/db/schema/zero-run";
 
@@ -14,6 +17,10 @@ import { clearMockNow, mockNow, now } from "../../../lib/time";
 import { server } from "../../../mocks/server";
 import { writeDb$ } from "../../external/db";
 import { seedAgentRunCallback$ } from "../../routes/__tests__/helpers/agent-run-callback";
+import {
+  captureGithubIssueApi,
+  mockGithubAppEnv,
+} from "../../routes/__tests__/helpers/api-bdd-github";
 import { createFixtureTracker } from "../../routes/__tests__/helpers/zero-route-test";
 import {
   deleteUsageInsightFixture$,
@@ -36,6 +43,8 @@ const TRIGGER_CRON_CALLBACK_URL =
   "http://localhost:3000/api/internal/callbacks/trigger/cron";
 const TRIGGER_LOOP_CALLBACK_URL =
   "http://localhost:3000/api/internal/callbacks/trigger/loop";
+const GITHUB_ISSUES_CALLBACK_URL =
+  "http://localhost:3000/api/internal/callbacks/github/issues";
 const OPENROUTER_COMPLETIONS_URL =
   "https://openrouter.ai/api/v1/chat/completions";
 
@@ -184,6 +193,68 @@ async function readTrigger(triggerId: string) {
     .where(eq(automationTriggers.id, triggerId))
     .limit(1);
   return row;
+}
+
+async function readGithubIssueSession(args: {
+  readonly installationId: string;
+  readonly repo: string;
+  readonly issueNumber: number;
+}) {
+  const db = store.set(writeDb$);
+  const [row] = await db
+    .select()
+    .from(githubIssueSessions)
+    .where(
+      and(
+        eq(githubIssueSessions.installationId, args.installationId),
+        eq(githubIssueSessions.repo, args.repo),
+        eq(githubIssueSessions.issueNumber, args.issueNumber),
+      ),
+    )
+    .limit(1);
+  return row;
+}
+
+async function seedGithubInstallation(
+  fixture: AgentCallbackRunFixture,
+): Promise<{
+  readonly installationId: string;
+  readonly remoteInstallationId: string;
+}> {
+  const db = store.set(writeDb$);
+  const numericRunId = fixture.runId.replace(/\D/gu, "").padEnd(9, "0");
+  const remoteInstallationId = `9${numericRunId.slice(0, 9)}`;
+  const [installation] = await db
+    .insert(githubInstallations)
+    .values({
+      installationId: remoteInstallationId,
+      status: "active",
+      orgId: fixture.orgId,
+      defaultComposeId: fixture.composeId,
+    })
+    .returning({ id: githubInstallations.id });
+  if (!installation) {
+    throw new Error("Failed to seed GitHub installation");
+  }
+  return { installationId: installation.id, remoteInstallationId };
+}
+
+async function seedGithubAgentSession(
+  fixture: AgentCallbackRunFixture,
+): Promise<string> {
+  const db = store.set(writeDb$);
+  const [session] = await db
+    .insert(agentSessions)
+    .values({
+      userId: fixture.userId,
+      orgId: fixture.orgId,
+      agentComposeId: fixture.composeId,
+    })
+    .returning({ id: agentSessions.id });
+  if (!session) {
+    throw new Error("Failed to seed GitHub agent session");
+  }
+  return session.id;
 }
 
 async function seedAutomationTrigger(
@@ -445,6 +516,207 @@ describe("dispatchRunCallbacks$ agent internal dispatch", () => {
         payload: { triggerAgentId: composeId },
       },
     ]);
+  });
+});
+
+describe("dispatchRunCallbacks$ GitHub issues internal dispatch", () => {
+  it("dispatches typed GitHub issue callbacks through ccstate without fetching the route", async () => {
+    mockGithubAppEnv();
+    const fixture = await seedAgentCallbackRun();
+    const installation = await seedGithubInstallation(fixture);
+    const newSessionId = await seedGithubAgentSession(fixture);
+    const issueApi = captureGithubIssueApi(installation.remoteInstallationId);
+    const { callbackId } = await store.set(
+      seedAgentRunCallback$,
+      {
+        runId: fixture.runId,
+        internalKind: "github:issues",
+        payload: {
+          installationId: installation.installationId,
+          repo: "vm0-ai/vm0",
+          issueNumber: 42,
+          agentId: fixture.composeId,
+          triggerCommentId: "1234",
+          triggerReactionId: "5678",
+          triggerCommentBody: "@Zero fix this",
+        },
+      },
+      context.signal,
+    );
+    mockRunOutput("GitHub callback finished.");
+    const routeRequests = failIfCallbackRouteIsFetched(
+      GITHUB_ISSUES_CALLBACK_URL,
+    );
+    const db = store.set(writeDb$);
+
+    const results = await store.set(
+      dispatchRunCallbacks$,
+      { db, runId: fixture.runId, status: "completed" },
+      context.signal,
+    );
+
+    expect(results).toStrictEqual([{ callbackId, success: true }]);
+    expect(routeRequests()).toBe(0);
+    expect(issueApi.comments).toHaveLength(1);
+    expect(issueApi.comments[0]).toMatchObject({
+      repo: "vm0-ai/vm0",
+      issueNumber: "42",
+    });
+    expect(issueApi.comments[0]?.body).toContain("> @Zero fix this");
+    expect(issueApi.comments[0]?.body).toContain("GitHub callback finished.");
+    expect(issueApi.reactionDeletes).toStrictEqual([
+      { commentId: "1234", reactionId: "5678" },
+    ]);
+    await expect(
+      readGithubIssueSession({
+        installationId: installation.installationId,
+        repo: "vm0-ai/vm0",
+        issueNumber: 42,
+      }),
+    ).resolves.toMatchObject({
+      agentSessionId: newSessionId,
+      lastCommentId: issueApi.comments[0]?.id,
+    });
+    await expect(readCallback(callbackId)).resolves.toMatchObject({
+      url: null,
+      internalKind: "github:issues",
+      status: "delivered",
+      attempts: 1,
+      lastError: null,
+    });
+  });
+
+  it("dispatches legacy GitHub issue callback URLs through the same ccstate path", async () => {
+    mockGithubAppEnv();
+    const fixture = await seedAgentCallbackRun();
+    const installation = await seedGithubInstallation(fixture);
+    const issueApi = captureGithubIssueApi(installation.remoteInstallationId);
+    const { callbackId } = await store.set(
+      seedAgentRunCallback$,
+      {
+        runId: fixture.runId,
+        url: GITHUB_ISSUES_CALLBACK_URL,
+        payload: {
+          installationId: installation.installationId,
+          repo: "vm0-ai/vm0",
+          issueNumber: 43,
+          agentId: fixture.composeId,
+        },
+      },
+      context.signal,
+    );
+    const routeRequests = failIfCallbackRouteIsFetched(
+      GITHUB_ISSUES_CALLBACK_URL,
+    );
+    const db = store.set(writeDb$);
+
+    const results = await store.set(
+      dispatchRunCallbacks$,
+      {
+        db,
+        runId: fixture.runId,
+        status: "failed",
+        error: "Legacy GitHub failure",
+      },
+      context.signal,
+    );
+
+    expect(results).toStrictEqual([{ callbackId, success: true }]);
+    expect(routeRequests()).toBe(0);
+    expect(issueApi.comments).toHaveLength(1);
+    expect(issueApi.comments[0]).toMatchObject({
+      repo: "vm0-ai/vm0",
+      issueNumber: "43",
+    });
+    expect(issueApi.comments[0]?.body).toContain("Oops, something went wrong.");
+    await expect(readCallback(callbackId)).resolves.toMatchObject({
+      url: GITHUB_ISSUES_CALLBACK_URL,
+      internalKind: null,
+      status: "delivered",
+      attempts: 1,
+      lastError: null,
+    });
+  });
+
+  it("marks invalid typed GitHub issue callback payloads as failed", async () => {
+    const fixture = await seedAgentCallbackRun();
+    const { callbackId } = await store.set(
+      seedAgentRunCallback$,
+      {
+        runId: fixture.runId,
+        internalKind: "github:issues",
+        payload: {},
+      },
+      context.signal,
+    );
+    const routeRequests = failIfCallbackRouteIsFetched(
+      GITHUB_ISSUES_CALLBACK_URL,
+    );
+    const db = store.set(writeDb$);
+
+    const results = await store.set(
+      dispatchRunCallbacks$,
+      { db, runId: fixture.runId, status: "completed" },
+      context.signal,
+    );
+
+    expect(results).toStrictEqual([
+      {
+        callbackId,
+        success: false,
+        error: "Invalid or missing payload",
+      },
+    ]);
+    expect(routeRequests()).toBe(0);
+    await expect(readCallback(callbackId)).resolves.toMatchObject({
+      status: "failed",
+      attempts: 1,
+      lastError: "Invalid or missing payload",
+    });
+  });
+
+  it("dispatches typed GitHub issue callbacks from the non-ccstate wrapper", async () => {
+    mockGithubAppEnv();
+    const fixture = await seedAgentCallbackRun();
+    const installation = await seedGithubInstallation(fixture);
+    const issueApi = captureGithubIssueApi(installation.remoteInstallationId);
+    const { callbackId } = await store.set(
+      seedAgentRunCallback$,
+      {
+        runId: fixture.runId,
+        internalKind: "github:issues",
+        payload: {
+          installationId: installation.installationId,
+          repo: "vm0-ai/vm0",
+          issueNumber: 44,
+          agentId: fixture.composeId,
+          sessionContinuityEnabled: false,
+        },
+      },
+      context.signal,
+    );
+    const routeRequests = failIfCallbackRouteIsFetched(
+      GITHUB_ISSUES_CALLBACK_URL,
+    );
+    const db = store.set(writeDb$);
+
+    const results = await dispatchRunCallbacks(
+      db,
+      fixture.runId,
+      "failed",
+      undefined,
+      "Non-ccstate GitHub failure",
+    );
+
+    expect(results).toStrictEqual([{ callbackId, success: true }]);
+    expect(routeRequests()).toBe(0);
+    expect(issueApi.comments).toHaveLength(1);
+    expect(issueApi.comments[0]?.body).toContain("Oops, something went wrong.");
+    await expect(readCallback(callbackId)).resolves.toMatchObject({
+      status: "delivered",
+      attempts: 1,
+      lastError: null,
+    });
   });
 });
